requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0

# LangChain + OpenAI Integration
langchain==0.2.14
langchain-openai==0.1.7
openai==1.40.0


# AI/ML Libraries
sentence-transformers >= 3.0.0
huggingface_hub >= 0.23.0

# File Processing
PyMuPDF==1.23.8  # fitz for PDF extraction
python-docx==1.1.0

# Additional utilities
python-multipart==0.0.6  # For file uploads in FastAPI
numpy==1.26.4
